+++ intern-book-app/src/app/components/card/card.component.html
@@ -1,15 +1,13 @@
-<div *ngIf="books">
-    <mat-card class="book-card">
-        <mat-card-title class="text-wrap">{{books.name}}</mat-card-title>
+    <mat-card class="book-card" *ngFor=" let book of booklist">
+        <mat-card-title class="text-wrap">{{book.name}}</mat-card-title>
         <mat-card-content class="text-wrap">
             <div>
-                {{books.detail}}
+                {{book.detail}}
             </div>
-            <div>評価:{{books.evaluation}}点</div>
+            <div>評価:{{book.evaluation}}点</div>
         </mat-card-content>
         <mat-card-actions align="end">
             <button mat-raised-button color="warn">削除</button>
         </mat-card-actions>
     </mat-card>
-
-</div>
+    <mat-card></mat-card>
